pyproject.toml
[project]
name = "fitnessbud"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.17.0",
    "asyncpg>=0.30.0",
    "asynctest>=0.13.0",
    "bcrypt>=5.0.0",
    "email-validator>=2.3.0",
    "fastapi>=0.119.0",
    "flask>=3.1.2",
    "gunicorn>=23.0.0",
    "jinja2>=3.1.6",
    "joblib>=1.5.2",
    "jwt>=1.4.0",
    "mock>=5.2.0",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "pandas-stubs==2.3.2.250926",
    "passlib>=1.7.4",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.11.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "scikit-learn>=1.7.2",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.37.0",
]
[tool.setuptools.packages]
# Explicitly list the folders that contain your Python code
find = { include = ["api", "core", "domain", "infrastructure"] }

[tool.pytest.ini_options]
# Set the asynchronous testing mode (necessary for pytest-asyncio)
asyncio_mode = "auto"
# Specify the directory where tests are located
testpaths = "tests"
# Standard naming convention for test files
python_files = "test_*.py"
